requirements.txt
ultralytics
streamlit
numpy
Pillow
opencv-python-headless
requests